add function to allow users to register their own parser hooks
A simple contrived example would be:

- banana.registerParserPlugin('foobar', nodes => nodes[0] === 'foo' ? nodes[1]: nodes[2])

Which makes
- banana.i18n('{{foobar:foo|first|second') --> 'first'
- banana.i18n('{{foobar:bar|first|second') --> 'second'

This can be used practically for a lot of purposes, particularly for operations that require some data to be available which can be fetched from the client script. For example on mediawiki sites, a client script could add a hook for formatting dates using the month names fetched from MediaWiki API.

Let me know what you think about this. Another less explicit way of allowing these hooks would be to export the BananaEmitter class, so that clients can directly modify its prototype.